pyproject.toml
[tool.poetry]
name = "django_main_muya_wce"
version = "1.0.0"
description = "MUYA-WCE Workspace for Collaborative Editing"
authors = ["Mary Chester-Kadwell <mec31@cam.ac.uk>"]
license = "GNU General Public License v3.0"

[tool.poetry.dependencies]
python = "^3.8"
celery = "^5.2.3"
Django = "^3.2"
django-celery-results = "^2.2.0"
django-environ = "^0.8.1"
psycopg2-binary = "2.9.4"
uWSGI = "^2.0.20"
backcall = "^0.2.0"
decorator = "^5.1.0"
djangorestframework = "^3.12.4"
jedi = "^0.18.0"
lxml = "^4.6.3"
matplotlib-inline = "^0.1.3"
pexpect = "^4.8.0"
pickleshare = "^0.7.5"
Pygments = "^2.10.0"
gevent = "^21.12.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
